Move Pragma class to new package
