+++ library/driver/src/commonMain/kotlin/io/toxicity/sqlite/mc/driver/SQLiteMCDriver.kt
@@ -19,7 +19,7 @@ import app.cash.sqldelight.db.QueryResult
 import app.cash.sqldelight.db.SqlSchema
 import io.toxicity.sqlite.mc.driver.config.*
 import io.toxicity.sqlite.mc.driver.config.encryption.*
-import io.toxicity.sqlite.mc.driver.config.mutablePragmas
+import io.toxicity.sqlite.mc.driver.config.pragma.mutablePragmas
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
 import kotlin.contracts.ExperimentalContracts
